cleaned up interfaces, moved relative files from interfaces into animal/habitat
