+++ interfaces/__init__.py
@@ -1,13 +1,13 @@
-from .terrestrial import ITerrestrial
-from .aquatic import IAquatic
-from .flying import IFlying
-from .walking import IWalking
-from .swimming import ISwimming
-from .saltwater import ISaltwater
-from .identifiable import Identifiable
-from .hospitable import IHospitable
-from .environmentCharacteristics import IEnviroChar
-from .habitat.stagnant import IStagnant
+from .animal.terrestrial import ITerrestrial
+from .animal.aquatic import IAquatic
+from .animal.flying import IFlying
+from .animal.walking import IWalking
+from .animal.swimming import ISwimming
+from .animal.saltwater import ISaltwater
 from .animal.freshwater import IFreshwater
+from .habitat.environmentCharacteristics import IEnviroChar
+from .habitat.stagnant import IStagnant
 from .habitat.contains_animals import IContainsAnimals
 from .habitat.contains_plants import IContainsPlants
+from .identifiable import Identifiable
+from .hospitable import IHospitable
